Add an overwrite protection during rendering.
Site generation starts with an empty dir. Then, a copy of static/ is
added. Then the rest of site is generated. If during generation we try
to write to a file that already exists, we're likely to be overwriting
a file from static/. Which is bad.

+++ py/render.py
@@ -66,6 +66,10 @@ class Renderer(object):
         except OSError, e:
             if e.errno == 17:
                 pass
+        if os.path.isfile(fn):
+            raise RuntimeError(
+                'Asked to generate a new page at %s. A file like that'
+                ' already exists. This should not happen.' % fn)
         with codecs.open(fn, 'w', 'utf-8') as f:
             if tmpl_fn:
                 template = self.jinja.get_template(tmpl_fn)
